加入OrderShippingInfo测试数据的准备样例.

+++ module/order/app/models/order/OrderShippingInfo.java
@@ -34,15 +34,17 @@ public class OrderShippingInfo extends Model {
 
     /**
      * 单价
+     * TODO: 删除这一属性，值放orderItems中了
      */
     @Column(name = "sale_price")
     public BigDecimal salePrice;
 
     /**
      * 购买数量
+     * TODO: 删除这一属性，值放orderItems中了
      */
     @Column(name = "buy_number")
-    public Integer buyNumber;
+    public Long buyNumber;
 
     /**
      * 备注
